first pass at connecting molecule counts to model

+++ simulations-java/simulations/acid-base-solutions/src/edu/colorado/phet/acidbasesolutions/model/concentration/BaseConcentrationModel.java
@@ -11,7 +11,7 @@ public abstract class BaseConcentrationModel extends ConcentrationModel {
 
     public abstract double getBaseConcentration();
     
-    public double getBaseMoleculeCount() {
+    public int getBaseMoleculeCount() {
         return getMoleculeCount( getBaseConcentration() );
     }
 }
